% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_url.R
\name{image_url}
\alias{image_url}
\title{Equivalent of `knitr::include_graphics()` for an image for these exercises}
\usage{
image_url(image_name)
}
\arguments{
\item{image_name}{Character string naming the image}
}
\description{
Equivalent of `knitr::include_graphics()` for an image for these exercises
}
